backend/pyproject.toml
[project]
name = "flexqr-backend"
version = "1.0.0"
description = "Layered FastAPI backend for the FlexQR open-source QR code management platform."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115,<1.0",
  "uvicorn[standard]>=0.30,<1.0",
  "pymongo>=4.10,<5.0",
  "pydantic-settings>=2.6,<3.0",
  "pyjwt>=2.9,<3.0",
  "email-validator>=2.2,<3.0",
  "google-auth>=2.36,<3.0",
  "python-dotenv>=1.0,<2.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
